+++ content/lib/pyswitch/clients/kemper/actions/effect_button.py
@@ -6,8 +6,8 @@ from ..mappings.effects import MAPPING_EFFECT_BUTTON
 from .. import KemperMappings
 
 # Effect Button I-IIII (set only). num must be a number (1 to 4).
-
-# NOTE: The state of the display will be toggled no matter what the real status on the Kemper is (the Kemper sadly does not report the state at all). 
+# 
+# <b>NOTE:</b> The state of the display will be toggled no matter what the real status on the Kemper is (the Kemper sadly does not report the state at all). 
 # If you dont want this behaviour, you can set "use_internal_state = False".
 def EFFECT_BUTTON(num,                           # Number of the Effect Button (1 to 4)
                   text = None,